backend/requirements.txt
# Core
fastapi>=0.115.0,<1.0.0
uvicorn>=0.34.0,<1.0.0
python-multipart>=0.0.20,<1.0.0
sqlalchemy>=2.0.0,<3.0.0

# Resume parsing
pdfplumber>=0.11.0,<1.0.0
python-docx>=1.1.0,<2.0.0

# LLM
ollama>=0.4.0,<1.0.0

# Config
python-dotenv>=1.0.0,<2.0.0
